magnav: Outsource tilt compensation to library
