+++ apps/waypointer/ChangeLog
@@ -5,3 +5,4 @@
 0.05: Fix not displaying of wpindex = 0
 0.06: Added adjustment for Bangle.js magnetometer heading fix
 0.07: Add settings file with the option to disable the slow direction updates
+0.08: Use tilt compensation from new magnav library
